refactor: modify theme-switch casing name
